Task #197 Dividir funciones lógicas del backend
Se crearon clases distintas para funcionalidades distintas en el backend
